Move simplates to their own submodule
